fix test and refactor test cases
